convert uint8 to uint256

+++ contracts/game/HeroManager.sol
@@ -24,7 +24,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
   uint256 public secondaryMultiplier = 8;
   uint256 public thirdMultiplier = 6;
 
-  uint8 public bonusExp = 30; // From Level 1, every battle win will give 30 exp to the hero. And as level goes up, this will be reduced. Level 1 -> 2: 30, Lv 2 -> 3: 29, ...., Lv 29 -> 30: 2
+  uint256 public bonusExp = 30; // From Level 1, every battle win will give 30 exp to the hero. And as level goes up, this will be reduced. Level 1 -> 2: 30, Lv 2 -> 3: 29, ...., Lv 29 -> 30: 2
 
   uint256 public rarityPowerBooster = 110;
 
@@ -42,7 +42,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     heroes[heroId] = hero;
   }
 
-  function levelUp(uint256 heroId, uint8 levels) public {
+  function levelUp(uint256 heroId, uint256 levels) public {
     require(nft.ownerOf(heroId) == msg.sender, "HeroManager: not a NFT owner");
     require(
       heroes[heroId].level < HERO_MAX_LEVEL,
@@ -127,7 +127,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     return heroes[heroId].level;
   }
 
-  function heroBonusExp(uint256 heroId) public view returns (uint8) {
+  function heroBonusExp(uint256 heroId) public view returns (uint256) {
     return bonusExp - heroes[heroId].level + 1;
   }
 
@@ -138,7 +138,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     );
 
     if (heroes[heroId].level < 30) {
-      uint8 gainedExp = heroBonusExp(heroId);
+      uint256 gainedExp = heroBonusExp(heroId);
       heroes[heroId].experience += gainedExp;
       if (heroes[heroId].experience >= 100) {
         heroes[heroId].experience -= 100;
@@ -165,7 +165,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     );
 
     if (heroes[heroId].level > 1 || heroes[heroId].experience > 0) {
-      uint8 gainedExp = heroBonusExp(heroId) / 2;
+      uint256 gainedExp = heroBonusExp(heroId) / 2;
       heroes[heroId].experience -= gainedExp;
       if (heroes[heroId].experience < 0) {
         if (heroes[heroId].level == 1) {
@@ -217,7 +217,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     bonusLevelUpFee = value;
   }
 
-  function setBonusExp(uint8 value) external onlyOwner {
+  function setBonusExp(uint256 value) external onlyOwner {
     bonusExp = value;
   }
